+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - Add test: logs error when primary and all fallback organization clones fail
 - Add tests for command add organization
 - add tests for url parser
+- Add tests for azure urls
 
 ### Changed